product category detail added

+++ app/Foundation/Interfaces/Http/Routes/admin/blog-categories.php
@@ -7,9 +7,9 @@ use Illuminate\Support\Facades\Route;
 
 
 Route::middleware(['auth:admin_api'])->prefix('admin')->group(function () {
-    Route::get('blog-categories', [BlogCategoriesController::class, 'categoryList']);
-    Route::get('blog-categories/{id}', [BlogCategoriesController::class, 'show']);
-    Route::post('blog-categories', [BlogCategoriesController::class, 'store']);
-    Route::put('blog-categories/{id}', [BlogCategoriesController::class, 'update']);
-    Route::delete('blog-categories/{id}', [BlogCategoriesController::class, 'destroy']);
+    Route::get('blog-categories', [BlogCategoriesController::class, 'categoryList'])->name('admin.blog-categories.list');
+    Route::get('blog-categories/{id}', [BlogCategoriesController::class, 'categoryShow'])->name('admin.blog-categories.show');
+    Route::post('blog-categories', [BlogCategoriesController::class, 'store'])->name('admin.blog-categories.store');
+    Route::put('blog-categories/{id}', [BlogCategoriesController::class, 'update'])->name('admin.blog-categories.update');
+    Route::delete('blog-categories/{id}', [BlogCategoriesController::class, 'destroy'])->name('admin.blog-categories.destroy');
 });
