+++ app/Foundation/Interfaces/Http/Routes/admin/blogs.php
@@ -7,9 +7,9 @@ use Illuminate\Support\Facades\Route;
 
 
 Route::middleware(['auth:admin_api'])->prefix('admin')->group(function () {
-    Route::get('blogs', [BlogsController::class, 'blogList']);
-    Route::get('blogs/{id}', [BlogsController::class, 'blogShow']);
-    Route::post('blogs', [BlogsController::class, 'blogStore']);
-    Route::put('blogs/{id}', [BlogsController::class, 'blogUpdate']);
-    Route::delete('blogs/{id}', [BlogsController::class, 'blogDestroy']);
+    Route::get('blogs', [BlogsController::class, 'blogList'])->name('admin.blogs.index');
+    Route::get('blogs/{id}', [BlogsController::class, 'blogShow'])->name('admin.blogs.show');
+    Route::post('blogs', [BlogsController::class, 'blogStore'])->name('admin.blogs.store');
+    Route::put('blogs/{id}', [BlogsController::class, 'blogUpdate'])->name('admin.blogs.update');
+    Route::delete('blogs/{id}', [BlogsController::class, 'blogDestroy'])->name('admin.blogs.destroy');
 });
